feat: compression for 2d textures and cubemaps. Can configure compression quality settings per-image now. The floor definitely has green artifacts in it though, more so when accounting for sRGB surprisingly. Artifacts in the skybox too, but less noticeable

+++ progression/graphics/graphics_api/texture.cpp
@@ -482,12 +482,12 @@ namespace Gfx
         return m_desc.format;
     }
 
-    uint8_t Texture::GetMipLevels() const
+    uint32_t Texture::GetMipLevels() const
     {
         return m_desc.mipLevels;
     }
 
-    uint8_t Texture::GetArrayLayers() const
+    uint32_t Texture::GetArrayLayers() const
     {
         return m_desc.arrayLayers;
     }
